Initial stab at factoring out an Editor wrapper class
