+++ src/DocumentManager.js
@@ -38,6 +38,7 @@ define(function (require, exports, module) {
         FileUtils           = require("FileUtils"),
         CommandManager      = require("CommandManager"),
         Async               = require("Async"),
+        Editor              = require("Editor").Editor,
         Commands            = require("Commands");
 
     /**
@@ -335,6 +336,8 @@ define(function (require, exports, module) {
      * @type {?CodeMirror}
      */
     Document.prototype._editor = null;
+    
+    Document.prototype.editor = null;
 
     /**
      * @private
@@ -356,23 +359,23 @@ define(function (require, exports, module) {
      * other than DocumentManager and EditorManager should access it.
      *
      * Initialize the editor instance for this file.
-     * @param {!CodeMirror} editor  The editor that will maintain the document state (current text
+     * @param {!Editor} editor  The editor that will maintain the document state (current text
      *          and undo stack). It is assumed that the editor text has already been initialized
-     *          with the file's contents. The editor may be null when the working set is restored
-     *          at initialization.
+     *          with the file's contents.
      * @param {!Date} initialTimestamp  Timestamp of file at the time we read its contents from disk.
      *          Required if editor is passed.
      * @perem {!string} rawText  Original text read from disk, beore handing to CodeMirror.
      */
     Document.prototype._setEditor = function (editor, initialTimestamp, rawText) {
         // Editor can only be assigned once per Document
-        console.assert(!this._editor);
+        console.assert(!this.editor);
         
-        this._editor = editor;
+        this.editor = editor;
+        this._editor = editor._editor;
         this.diskTimestamp = initialTimestamp;
         
         // Dirty-bit tracking
-        editor.setOption("onChange", this._handleEditorChange.bind(this));
+        $(editor).on("onChange", this._handleEditorChange.bind(this));
         this.isDirty = false;
         
         // Sniff line-ending style
@@ -424,7 +427,7 @@ define(function (require, exports, module) {
     Document.prototype.getText = function () {
         // CodeMirror.getValue() always returns text with LF line endings; fix up to match line
         // endings preferred by the document, if necessary
-        var codeMirrorText = this._editor.getValue();
+        var codeMirrorText = this.editor.getText();
         if (this._lineEndings === FileUtils.LINE_ENDINGS_LF) {
             return codeMirrorText;
         } else {
@@ -438,7 +441,7 @@ define(function (require, exports, module) {
      * @param {!string} text The text to replace the contents of the document with.
      */
     Document.prototype.setText = function (text) {
-        this._editor.setValue(text);
+        this.editor.setText(text);
     };
     
     /**
@@ -449,8 +452,7 @@ define(function (require, exports, module) {
      * @param {!Date} newTimestamp Timestamp of file at the time we read its new contents from disk.
      */
     Document.prototype.refreshText = function (text, newTimestamp) {
-        this._editor.setValue(text);
-        this._editor.clearHistory();
+        this.editor.resetText(text);
         
         this._markClean();
         this.diskTimestamp = newTimestamp;
@@ -463,22 +465,9 @@ define(function (require, exports, module) {
     };
     
     /**
-     * Sets the cursor position in the document.
-     * @param {number} line The 0 based line number.
-     * @param {number} char The 0 based character position.
-     */
-    Document.prototype.setCursor = function (line, char) {
-        this._editor.setCursor(line, char);
-    };
-    
-    /**
      * @private
      */
     Document.prototype._handleEditorChange = function () {
-        if (!this._editor) {
-            return;
-        }
-
         // On any change, mark the file dirty. In the future, we should make it so that if you
         // undo back to the last saved state, we mark the file clean.
         var wasDirty = this.isDirty;
@@ -495,7 +484,7 @@ define(function (require, exports, module) {
      * @private
      */
     Document.prototype._markClean = function () {
-        if (!this._editor) {
+        if (!this.editor) {
             return;
         }
