+++ src/EditorManager.js
@@ -25,6 +25,7 @@ define(function (require, exports, module) {
     // Load dependent modules
     var FileUtils           = require("FileUtils"),
         DocumentManager     = require("DocumentManager"),
+        Editor              = require("Editor").Editor,
         EditorUtils         = require("EditorUtils"),
         Strings             = require("strings");
     
@@ -46,118 +47,6 @@ define(function (require, exports, module) {
     var _inlineEditProviders = [];
     
     /**
-     * @private
-     * Handle Tab key press.
-     * @param {!CodeMirror} instance CodeMirror instance.
-     */
-    function _handleTabKey(instance) {
-        // Tab key handling is done as follows:
-        // 1. If the selection is before any text and the indentation is to the left of 
-        //    the proper indentation then indent it to the proper place. Otherwise,
-        //    add another tab. In either case, move the insertion point to the 
-        //    beginning of the text.
-        // 2. If the selection is after the first non-space character, and is not an 
-        //    insertion point, indent the entire line(s).
-        // 3. If the selection is after the first non-space character, and is an 
-        //    insertion point, insert a tab character or the appropriate number 
-        //    of spaces to pad to the nearest tab boundary.
-        var from = instance.getCursor(true),
-            to = instance.getCursor(false),
-            line = instance.getLine(from.line),
-            indentAuto = false,
-            insertTab = false;
-        
-        if (from.line === to.line) {
-            if (line.search(/\S/) > to.ch || to.ch === 0) {
-                indentAuto = true;
-            }
-        }
-
-        if (indentAuto) {
-            var currentLength = line.length;
-            CodeMirror.commands.indentAuto(instance);
-            // If the amount of whitespace didn't change, insert another tab
-            if (instance.getLine(from.line).length === currentLength) {
-                insertTab = true;
-                to.ch = 0;
-            }
-        } else if (instance.somethingSelected()) {
-            CodeMirror.commands.indentMore(instance);
-        } else {
-            insertTab = true;
-        }
-        
-        if (insertTab) {
-            if (instance.getOption("indentWithTabs")) {
-                CodeMirror.commands.insertTab(instance);
-            } else {
-                var i, ins = "", numSpaces = instance.getOption("tabSize");
-                numSpaces -= to.ch % numSpaces;
-                for (i = 0; i < numSpaces; i++) {
-                    ins += " ";
-                }
-                instance.replaceSelection(ins, "end");
-            }
-        }
-    }
-    
-    /**
-     * @private
-     * Handle left arrow, right arrow, backspace and delete keys when soft tabs are used.
-     * @param {!CodeMirror} instance CodeMirror instance 
-     * @param {number} direction Direction of movement: 1 for forward, -1 for backward
-     * @param {function} functionName name of the CodeMirror function to call
-     * @return {boolean} true if key was handled
-     */
-    function _handleSoftTabNavigation(instance, direction, functionName) {
-        var handled = false;
-        if (!instance.getOption("indentWithTabs")) {
-            var cursor = instance.getCursor(),
-                tabSize = instance.getOption("tabSize"),
-                jump = cursor.ch % tabSize,
-                line = instance.getLine(cursor.line);
-
-            if (direction === 1) {
-                jump = tabSize - jump;
-
-                if (cursor.ch + jump > line.length) { // Jump would go beyond current line
-                    return false;
-                }
-
-                if (line.substr(cursor.ch, jump).search(/\S/) === -1) {
-                    instance[functionName](jump, "char");
-                    handled = true;
-                }
-            } else {
-                // Quick exit if we are at the beginning of the line
-                if (cursor.ch === 0) {
-                    return false;
-                }
-                
-                // If we are on the tab boundary, jump by the full amount, 
-                // but not beyond the start of the line.
-                if (jump === 0) {
-                    jump = tabSize;
-                }
-
-                // Search backwards to the first non-space character
-                var offset = line.substr(cursor.ch - jump, jump).search(/\s*$/g);
-
-                if (offset !== -1) { // Adjust to jump to first non-space character
-                    jump -= offset;
-                }
-
-                if (jump > 0) {
-                    instance[functionName](-jump, "char");
-                    handled = true;
-                }
-            }
-        }
-
-        return handled;
-    }
-    
-    /**
      * Creates a new CodeMirror editor instance containing the given text. The editor's mode is set
      * based on the given filename's extension (the actual file on disk is never examined). The
      * editor is not yet visible.
@@ -167,88 +56,21 @@ define(function (require, exports, module) {
      * @param {!jQueryObject} container  Container to add the editor to.
      * @param {!function} onInlineGesture  Handler for Ctrl+E command (open/close inline, depending
                 on context)
-     * @return {CodeMirror} the newly created editor.
+     * @return {Editor} the newly created editor.
      */
     function _createEditorFromText(text, fileNameToSelectMode, container, onInlineGesture) {
-        // NOTE: CodeMirror doesn't actually require calling 'new',
-        // but jslint does require it because of the capital 'C'
-        var editor = new CodeMirror(container, {
-            electricChars: false,
-            indentUnit : 4,
-            lineNumbers: true,
-            matchBrackets: true,
-            extraKeys: {
-                "Tab"  : _handleTabKey,
-                "Left" : function (instance) {
-                    if (!_handleSoftTabNavigation(instance, -1, "moveH")) {
-                        CodeMirror.commands.goCharLeft(instance);
-                    }
-                },
-                "Right" : function (instance) {
-                    if (!_handleSoftTabNavigation(instance, 1, "moveH")) {
-                        CodeMirror.commands.goCharRight(instance);
-                    }
-                },
-                "Backspace" : function (instance) {
-                    if (!_handleSoftTabNavigation(instance, -1, "deleteH")) {
-                        CodeMirror.commands.delCharLeft(instance);
-                    }
-                },
-                "Delete" : function (instance) {
-                    if (!_handleSoftTabNavigation(instance, 1, "deleteH")) {
-                        CodeMirror.commands.delCharRight(instance);
-                    }
-                },
-                "F3": "findNext",
-                "Shift-F3": "findPrev",
-                "Ctrl-H": "replace",
-                "Shift-Delete": "cut",
-                "Ctrl-Insert": "copy",
-                "Shift-Insert": "paste",
-                "Ctrl-E" : function (instance) {
-                    onInlineGesture(instance);
-                },
-                "Cmd-E" : function (instance) {
-                    onInlineGesture(instance);
-                }
+        var mode = EditorUtils.getModeFromFileExtension(fileNameToSelectMode);
+        
+        var extraKeys = {
+            "Ctrl-E" : function (editor) {
+                onInlineGesture(editor);
             },
-            onKeyEvent: function (instance, event) {
-                if (event.type === "keypress") {
-                    var keyStr = String.fromCharCode(event.which || event.keyCode);
-                    if (/[\]\}\)]/.test(keyStr)) {
-                        // If the whole line is whitespace, auto-indent it
-                        var lineNum = instance.getCursor().line;
-                        var lineStr = instance.getLine(lineNum);
-                        
-                        if (!/\S/.test(lineStr)) {
-                            // Need to do the auto-indent on a timeout to ensure
-                            // the keypress is handled before auto-indenting.
-                            // This is the same timeout value used by the
-                            // electricChars feature in CodeMirror.
-                            setTimeout(function () {
-                                instance.indentLine(lineNum);
-                            }, 75);
-                        }
-                    }
-                }
-                
-                
-                $(exports).triggerHandler("onKeyEvent", [instance, event]);
-                return false;
+            "Cmd-E" : function (editor) {
+                onInlineGesture(editor);
             }
-        });
-        
-        // Set code-coloring mode
-        EditorUtils.setModeFromFileExtension(editor, fileNameToSelectMode);
-        
-        // Initially populate with text. This will send a spurious change event, but that's ok
-        // because no one's listening yet (and we clear the undo stack below)
-        editor.setValue(text);
-        
-        // Make sure we can't undo back to the empty state before setValue()
-        editor.clearHistory();
-        
-        return editor;
+        };
+
+        return new Editor(text, mode, container, extraKeys);
     }
     
     /**
@@ -269,20 +91,20 @@ define(function (require, exports, module) {
     }
     
     /** Bound to Ctrl+E on outermost editors */
-    function _openInlineWidget(instance) {
+    function _openInlineWidget(editor) {
         // Run through inline-editor providers until one responds
-        var pos = instance.getCursor();
+        var pos = editor.getCursorPos();
         var inlinePromise;
         var i;
         for (i = 0; i < _inlineEditProviders.length && !inlinePromise; i++) {
             var provider = _inlineEditProviders[i];
-            inlinePromise = provider(instance, pos);
+            inlinePromise = provider(editor._editor, pos);
         }
         
         // If one of them will provide a widget, show it inline once ready
         if (inlinePromise) {
             inlinePromise.done(function (inlineContent) {
-                var inlineId = instance.addInlineWidget(pos, inlineContent.content, inlineContent.height);
+                var inlineId = editor._editor.addInlineWidget(pos, inlineContent.content, inlineContent.height);
                 inlineContent.onAdded(inlineId);
             });
         }
@@ -365,8 +187,8 @@ define(function (require, exports, module) {
             reader = FileUtils.readAsText(fileEntry);
             
         reader.done(function (text, readTimestamp) {
-            var editor = _createEditorFromText(text, fileEntry.fullPath, container, _openInlineWidget);
-            result.resolve(editor, readTimestamp, text);
+            var editorWrapper = _createEditorFromText(text, fileEntry.fullPath, container, _openInlineWidget);
+            result.resolve(editorWrapper, readTimestamp, text);
         });
         reader.fail(function (error) {
             result.reject(error);
@@ -393,16 +215,17 @@ define(function (require, exports, module) {
         $(inlineContent).addClass("inlineCodeEditor");
         
         var myInlineId; // won't be populated until our afterAdded() callback is run
-        function closeThisInline(instance) {
+        function closeThisInline(editor) {
             _closeInlineWidget(hostEditor, myInlineId);
             var doc = _getDocumentForEditor(hostEditor);
             if (doc) {
-                doc._removeInlineEditor(instance);
+                doc._removeInlineEditor(editor._editor);
             }
             _syncGutterWidths(hostEditor);
         }
         
-        var inlineEditor = _createEditorFromText(text, fileNameToSelectMode, inlineContent, closeThisInline);
+        var editorWrapper = _createEditorFromText(text, fileNameToSelectMode, inlineContent, closeThisInline);
+        var inlineEditor = editorWrapper._editor;
         
         // Some tasks have to wait until we've been parented into the outer editor
         function afterAdded(inlineId) {
@@ -481,6 +304,7 @@ define(function (require, exports, module) {
     }
 
     /** Focus the currently visible editor. If no editor visible, does nothing. */
+    // TODO: rename to focusMainEditor()
     function focusEditor() {
         if (_currentEditor) {
             _currentEditor.focus();
@@ -543,11 +367,11 @@ define(function (require, exports, module) {
         }
 
         // Lazily create editor for Documents that were restored on-init
-        if (!document._editor) {
+        if (!document.editor) {
             var editorResult = _createEditorFromFile(document.file, _editorHolder.get(0));
 
-            editorResult.done(function (editor, readTimestamp, rawText) {
-                document._setEditor(editor, readTimestamp, rawText);
+            editorResult.done(function (editorWrapper, readTimestamp, rawText) {
+                document._setEditor(editorWrapper, readTimestamp, rawText);
                 _doShow(document);
             });
             editorResult.fail(function (error) {
@@ -638,10 +462,10 @@ define(function (require, exports, module) {
         var result          = new $.Deferred(),
             editorResult    = _createEditorFromFile(fileEntry, _editorHolder.get(0));
 
-        editorResult.done(function (editor, readTimestamp, rawText) {
+        editorResult.done(function (editorWrapper, readTimestamp, rawText) {
             // Create the Document wrapping editor & binding it to a file
             var doc = new DocumentManager.Document(fileEntry);
-            doc._setEditor(editor, readTimestamp, rawText);
+            doc._setEditor(editorWrapper, readTimestamp, rawText);
             result.resolve(doc);
         });
