+++ src/index.html
@@ -27,9 +27,6 @@
 
 </head>
 <body>
-    <!-- TODO (Issue #279): LESS won't work directly from Chrome when run locally (unless you launch it with -allow-file-access-from-files)
-    due to annoying security restrictions. This will work in the shell app, or in other browsers like Safari. -->
-
     <!-- Top bar for menus -->
     <div class="topbar" data-dropdown="dropdown">
         <div class="topbar-inner">
